browtrix/apps/server/pyproject.toml
[project]
name = "browtrix-server"
version = "0.1.0"
description = "Browtrix MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "pydantic-settings>=2.0.0",
    "aiohttp>=3.13.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "structlog>=24.0.0",
]

[project.scripts]
browtrix-server = "browtrix_server.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ty>=0.0.1a27",
    "ruff>=0.6.0",
]
